repo-updater: Move GitoliteSource to gitolite.go
